# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Tài liệu API",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tạo người dùng mới",
                "description": "API này cho phép tạo người dùng mới.",
                "operationId": "createUser",
                "responses": {
                    "201": {
                        "description": "Thông tin người dùng mới",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "tạo người dùng mới thành công"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "role_id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "leanhuynh"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "leanhuynh2002@gmail.com"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "http:/avatar.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server bị lỗi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "servers bị lỗi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/update": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Cập nhật thông tin người dùng",
                "description": "API này cho phép cập nhật thông tin người dùng.",
                "operationId": "updateUser",
                "responses": {
                    "200": {
                        "description": "Thông tin người dùng mới cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "cập nhật thông tin người dùng thành công"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "role_id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "leanhuynh"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "leanhuynh2002@gmail.com"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "http:/avatar.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/delete": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Xóa người dùng theo id",
                "description": "API này cho phép xóa người dùng theo id.",
                "operationId": "deleteUser",
                "responses": {
                    "200": {
                        "description": "Thông báo xóa người dùng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa tài khoản thành công"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "role_id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "leanhuynh"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "leanhuynh2002@gmail.com"
                                                },
                                                "avatar": {
                                                    "type": "string",
                                                    "example": "http:/avatar.com"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/search": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Tìm kiếm người dùng theo từ khóa",
                "description": "API này cho phép tìm kiếm người dùng theo từ khóa trong tên hoặc email.",
                "operationId": "searchUsers",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm người dùng",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách người dùng theo name hoặc email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "role_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "leanhuynh"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "leanhuynh2002@gmail.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "32$!sdsdf%3245"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "http://avatar.com"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Lấy danh sách người dùng",
                "description": "Trả về danh sách người dùng",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "View với danh sách người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "<html>...</html>"
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}